refactored MPSC
